+++ CHANGELOG.md
@@ -74,6 +74,7 @@ Changes since v2.4
 - Email messages now use the application title and full names of users
 - Email message texts improved
 - Show three latest events as a default on the application page (#1632)
+- Event descriptions on application page now use full name and are more thorough (#1634)
 
 ### Fixes
 - Entitlement API